+++ redfishMockupServer.py
@@ -9,10 +9,11 @@ import re
 import sys
 import argparse
 import time
-import collections
+import collections.abc
 import json
 import threading
 import datetime
+import signal
 
 import grequests
 
@@ -29,7 +30,7 @@ ch = logging.StreamHandler(sys.stdout)
 ch.setLevel(logging.INFO)
 logger.addHandler(ch)
 
-tool_version = "1.1.8"
+tool_version = "1.2.1"
 
 dont_send = ["connection", "keep-alive", "content-length", "transfer-encoding"]
 
@@ -46,7 +47,7 @@ def dict_merge(dct, merge_dct):
         :return: None
         """
         for k in merge_dct:
-            if (k in dct and isinstance(dct[k], dict) and isinstance(merge_dct[k], collections.Mapping)):
+            if (k in dct and isinstance(dct[k], dict) and isinstance(merge_dct[k], collections.abc.Mapping)):
                 dict_merge(dct[k], merge_dct[k])
             else:
                 dct[k] = merge_dct[k]
@@ -405,7 +406,7 @@ class RfMockupServer(BaseHTTPRequestHandler):
                         filename, file_extension = os.path.splitext(fpath_direct)
                         file_extension = file_extension.strip('.')
                     self.send_response(200)
-                    self.send_header("Content-Type", "application/" + file_extension + ";odata.metadata=minimal;charset=utf-8")
+                    self.send_header("Content-Type", "application/" + file_extension + ";charset=utf-8")
                     self.send_header("OData-Version", "4.0")
                 else:
                     self.send_response(404)
@@ -473,7 +474,7 @@ class RfMockupServer(BaseHTTPRequestHandler):
                 # Query Subscriptions should not return HttpHeaders.
                 if 'EventService/Subscriptions' in self.path:
                    if output_data.get('HttpHeaders') is not None:
-                      # This array is null or an empty array in responses. 
+                      # This array is null or an empty array in responses.
                       # An empty array is the preferred return value on read operations.
                       output_data['HttpHeaders'] = []
 
@@ -510,24 +511,44 @@ class RfMockupServer(BaseHTTPRequestHandler):
                 if not (self.server.headers and (os.path.isfile(fpath_headers))):
                     self.send_header("Content-Length", len(encoded_data))
                 self.end_headers()
-                
+
                 self.wfile.write(encoded_data)
 
             # if XML...
-            elif(os.path.isfile(fpath_xml) or os.path.isfile(fpath_direct)):
-                if os.path.isfile(fpath_xml):
-                    file_extension = 'xml'
-                    f = open(fpath_xml, "r")
-                elif os.path.isfile(fpath_direct):
-                    filename, file_extension = os.path.splitext(fpath_direct)
-                    file_extension = file_extension.strip('.')
-                    f = open(fpath_direct, "r")
+            elif os.path.isfile(fpath_xml):
+                f = open(fpath_xml, "r")
                 self.send_response(200)
-                self.send_header("Content-Type", "application/" + file_extension + ";odata.metadata=minimal;charset=utf-8")
+                self.send_header("Content-Type", "application/xml;charset=utf-8")
                 self.send_header("OData-Version", "4.0")
                 self.end_headers()
                 self.wfile.write(f.read().encode())
                 f.close()
+
+            elif os.path.isfile(fpath_direct):
+                self.send_response(200)
+                with open(fpath_direct, 'rb') as f:
+                    content = f.read()
+
+                try:
+                    decoded_content = content.decode()
+                except ValueError:
+                    # If the file is binary, send it as octet-stream.
+                    self.send_header("Content-Type", "application/octet-stream")
+                    self.send_header("OData-Version", "4.0")
+                    self.end_headers()
+                    self.wfile.write(content)
+                else:
+                    # The file is text.
+                    file_extension = os.path.splitext(fpath_direct)[1]
+                    if file_extension == "":
+                        mime_type = "text/plain"
+                    else:
+                        mime_type = "application/" + file_extension[1:]
+                    self.send_header("Content-Type", mime_type + ";charset=utf-8")
+                    self.send_header("OData-Version", "4.0")
+                    self.end_headers()
+                    self.wfile.write(decoded_content.encode("utf-8"))
+
             else:
                 self.send_response(404)
                 self.end_headers()
@@ -822,6 +843,12 @@ def main():
                 sys.exit(1)
 
         myServer = HTTPServer((hostname, port), RfMockupServer)
+
+        def sigterm_handler(signal_number, frame):
+            logger.info("SIGTERM: Shutting down http server")
+            myServer.server_close()
+            sys.exit(0)
+        signal.signal(signal.SIGTERM, sigterm_handler)
 
         if sslMode:
             logger.info("Using SSL with certfile: {}".format(sslCert))